+++ react/InputFieldContainer.js
@@ -28,7 +28,7 @@ class InputFieldContainer extends Component {
           }
           : cleanAddress
       },
-      {}
+      {},
     )
   }
 
@@ -57,10 +57,9 @@ class InputFieldContainer extends Component {
 
     return () => {
       onChangeAddress({
-        'number': {
+        number: {
           ...address['number'],
-          value: address['number'].disabled
-            ? null : labelNotApplicable,
+          value: address['number'].disabled ? null : labelNotApplicable,
           disabled: !address['number'].disabled,
         },
       })
@@ -121,6 +120,8 @@ class InputFieldContainer extends Component {
       autoFocus,
       address,
       options,
+      onSubmit,
+      submitLabel,
       rules,
       shouldShowNumberKeyboard,
     } = this.props
@@ -137,12 +138,14 @@ class InputFieldContainer extends Component {
         field={field}
         autoFocus={autoFocus}
         options={_options}
+        onSubmit={onSubmit}
+        submitLabel={submitLabel}
         onChange={this.bindOnChange()}
         onBlur={this.bindOnBlur()}
-        {...(address[field.name].notApplicable
+        {...(address[field.name].notApplicable &&
+        address[field.name].notApplicable
           ? { toggleNotApplicable: this.bindNotApplicable() }
-          : {}
-        )}
+          : {})}
         inputRef={this.inputRef}
         shouldShowNumberKeyboard={shouldShowNumberKeyboard}
       />
@@ -163,8 +166,10 @@ InputFieldContainer.propTypes = {
   rules: PropTypes.object.isRequired,
   options: PropTypes.array,
   onChangeAddress: PropTypes.func.isRequired,
-  shouldShowNumberKeyboard: PropTypes.bool,
+  onSubmit: PropTypes.func,
   notApplicableLabel: PropTypes.string,
+  shouldShowNumberKeyboard: PropTypes.bool,
+  submitLabel: PropTypes.string,
 }
 
 export default pureInputField(InputFieldContainer)