+++ react/PostalCodeGetter.js
@@ -11,7 +11,7 @@ import { getField } from './selectors/fields'
 import { injectRules } from './addressRulesContext'
 import { compose } from 'recompose'
 import { injectAddressContext } from './addressContainerContext'
-import { injectIntl } from 'react-intl'
+import { injectIntl, intlShape } from 'react-intl'
 
 class PostalCodeGetter extends Component {
   render() {
@@ -22,6 +22,8 @@ class PostalCodeGetter extends Component {
       onChangeAddress,
       Input,
       intl,
+      onSubmit,
+      submitLabel,
       shouldShowNumberKeyboard,
     } = this.props
 
@@ -34,6 +36,8 @@ class PostalCodeGetter extends Component {
             address={address}
             rules={rules}
             onChangeAddress={onChangeAddress}
+            onSubmit={onSubmit}
+            submitLabel={submitLabel}
           />
         )
       case TWO_LEVELS:
@@ -44,6 +48,8 @@ class PostalCodeGetter extends Component {
             address={address}
             rules={rules}
             onChangeAddress={onChangeAddress}
+            onSubmit={onSubmit}
+            submitLabel={submitLabel}
           />
         )
       case ONE_LEVEL:
@@ -54,6 +60,8 @@ class PostalCodeGetter extends Component {
             address={address}
             rules={rules}
             onChangeAddress={onChangeAddress}
+            onSubmit={onSubmit}
+            submitLabel={submitLabel}
           />
         )
       default:
@@ -68,6 +76,8 @@ class PostalCodeGetter extends Component {
             autoFocus={autoFocus}
             rules={rules}
             onChangeAddress={onChangeAddress}
+            onSubmit={onSubmit}
+            submitLabel={submitLabel}
             shouldShowNumberKeyboard={shouldShowNumberKeyboard}
           />
         )
@@ -83,12 +93,15 @@ PostalCodeGetter.defaultProps = {
 }
 
 PostalCodeGetter.propTypes = {
-  Input: PropTypes.func,
-  autoFocus: PropTypes.bool,
   address: AddressShapeWithValidation,
-  rules: PropTypes.object.isRequired,
+  autoFocus: PropTypes.bool,
+  Input: PropTypes.func,
+  intl: intlShape.isRequired,
   onChangeAddress: PropTypes.func.isRequired,
+  onSubmit: PropTypes.func,
+  rules: PropTypes.object.isRequired,
   shouldShowNumberKeyboard: PropTypes.bool,
+  submitLabel: PropTypes.string,
 }
 
 const enhance = compose(